requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
click==8.1.7
docx==0.2.4
Flask==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
packaging==24.0
pillow==10.3.0
pytesseract==0.3.10
python-docx==1.1.2
soupsieve==2.5
tesserocr==2.7.0
typing_extensions==4.12.0
Werkzeug==3.0.3
